Stop re-creating ConversationInputToolbar on theme change.
I've noticed that iOS does change theme multiple times when the app is backgrounded (probably creating screenshots for the app switcher). And each change caused `ConversationInputToolbar` to be re-created and replaced - super ineffective and potentially? causing UI glitches.

Recent changes made to support iOS 26 have already added support for adjusting to dark/light UI style on the fly. What was needed to be done is extending that functionality to support UI used on iOS 15-18, which these changes do.