Workaround concurrency issues in conversation view controller
This does not fully resolve the concurrency issues with conversation
view cell reuse and the edit UI being accessible on rows it should not
be available on but this method in EditManagerImpl appears to be
backstopping the other message types that are not editable. Making
polls uneditable as well in this method makes it so the edit UI cannot
be entered and that appears to improve the situation over in the
conversation view.